Parse transform implemented
The parse transform allows to avoid including the hamcrest header
into the test module. Including the module required to have the
namespaced_types define to be set up in the application using this
library. That was inconvenient.

The asserts generated by the parse transform avoids temporary
functions in the stack trace of the failed assert. This makes
it more convenient to use in CT.